Rewrite and integrate data2aug
