+++ .gitignore
@@ -1,2 +1,3 @@
 .idea
-samples2
+samples
+df_generated
