+++ README.md
@@ -1,3 +1,3 @@
-Aug2coco is form [Aug2Coco-converter](https://github.com/grrvk/Aug2Coco-converter.git).
+Aug2coco is from [Aug2Coco-converter](https://github.com/grrvk/Aug2Coco-converter.git).
 
-Data2aug is form [Data-gen](https://github.com/grrvk/Data-gen.git).
+Data2aug is from [Data-gen](https://github.com/grrvk/Data-gen.git).
