+++ data2aug/README.md
@@ -1,0 +1,6 @@
+imgaug transformation depends on imagecorruptions.
+To install imagecorruptions run:
+
+```bash
+pip install imagecorruptions
+```
